pyprojects/secretnote/pyproject.toml
[project]
authors = [
  {name = "wenyu.jqq", email = "wenyu.jqq@antgroup.com"},
  {name = "Tony Wu", email = "tonywu6@protonmail.com"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: Apache Software License",
]
dependencies = [
  "click>=8.1.7",
  "fastapi<100.0.0",
  "importlib-resources>=5.12",
  "ipykernel>=6.24.0",
  "ipython-sql~=0.5",
  "ipywidgets>=8.1.1",
  "jinja2>=3.1.2",
  "jupyter-resource-usage~=1.0",
  "jupyter-server~=2.7",
  "loguru>=0.7.2",
  "more-itertools>=10.1.0",
  "orjson>=3.9.9",
  "packaging>=23.1",
  "pydantic>=1.10, <2",
  "pyhumps>=3.8.0",
  "rich>=13.7.0",
  "tqdm>=4.66.1",
]
description = "Notebook suite for SecretFlow"
dynamic = ["version"]
license = "Apache-2.0"
name = "secretnote"
readme = "README.md"
requires-python = ">=3.8, <3.12"

[project.optional-dependencies]
docpack = [
  "beautifulsoup4>=4.12.2",
  "lxml>=4.9.3",
  "pillow>=8.0.0",
]
instrumentation = [
  "astunparse>=1.6.3",
  "networkx>=2.8.8",
  "opentelemetry-api>=1.20.0",
  "opentelemetry-sdk>=1.20.0",
  "ray>=2.2.0",
  "stack_data>=0.6.3",
]
secretflow = [
  "interconnection>=0.1.0.dev20231204",
  "spu>=0.6.0b0",
  "sf-heu>=0.4.4b0",
]

[project.scripts]
secretnote = "secretnote.server.app:SecretNoteApp.launch"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "packaging>=23.1"]

[tool.rye]
dev-dependencies = [
  "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
  "ray[default]~=2.6.3",
  "secretflow-lite==1.3.0.dev20231219",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
expression = "__version__"
path = "./src/secretnote/__init__.py"
search-paths = ["./src"]
source = "code"

[tool.hatch.build]
artifacts = ["src/secretnote/**/dist/**/*", "*.js", "*.cjs"]
directory = "dist"
only-include = ["src", "package.json"]

[tool.hatch.build.targets.wheel]
sources = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ruff.extend-per-file-ignores]
"src/secretnote/proto/**/*" = ["E501", "I001"]
